Refactor numeric literal handling in the `parser` package
- Move the `splitNumericLiteral`, `isDigitForBase`, `isValidIntegerSuffix`, and `isLetter` functions from `types.go` to a new `types_utils.go` file for better organization and separation of concerns.
- Enhance the `splitNumericLiteral` function to handle various numeric bases (binary, octal, decimal, hexadecimal) and improve error reporting for invalid literals.
- Introduce utility functions to validate integer suffixes and check for letter characters, aiding in the parsing process.

This commit improves the structure of the parser's utility functions, facilitating easier maintenance and enhancing the clarity of numeric literal parsing in the Surge language.
